+++ lib/drop_cap_text.dart
@@ -1,9 +1,6 @@
 import 'dart:math';
-import 'dart:ui';
 
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/rendering.dart';
 
 enum DropCapMode {
   /// default
@@ -26,13 +23,11 @@ class DropCap extends StatelessWidget {
   final double width, height;
 
   DropCap({
-    Key? key,
+    super.key,
     required this.child,
     required this.width,
     required this.height,
-  })   : assert(width != null),
-        assert(height != null),
-        super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -55,52 +50,61 @@ class DropCapText extends StatelessWidget {
   final int? maxLines;
   final TextOverflow overflow;
 
-  DropCapText(this.data,
-      {Key? key,
-      this.mode = DropCapMode.inside,
-      this.style,
-      this.dropCapStyle,
-      this.textAlign = TextAlign.start,
-      this.dropCap,
-      this.dropCapPadding = EdgeInsets.zero,
-      this.indentation = Offset.zero,
-      this.dropCapChars = 1,
-      this.forceNoDescent = false,
-      this.parseInlineMarkdown = false,
-      this.textDirection = TextDirection.ltr,
-      this.overflow = TextOverflow.clip,
-      this.maxLines,
-      this.dropCapPosition})
-      : assert(data != null),
-        super(key: key);
+  DropCapText(
+    this.data, {
+    super.key,
+    this.mode = DropCapMode.inside,
+    this.style,
+    this.dropCapStyle,
+    this.textAlign = TextAlign.start,
+    this.dropCap,
+    this.dropCapPadding = EdgeInsets.zero,
+    this.indentation = Offset.zero,
+    this.dropCapChars = 1,
+    this.forceNoDescent = false,
+    this.parseInlineMarkdown = false,
+    this.textDirection = TextDirection.ltr,
+    this.overflow = TextOverflow.clip,
+    this.maxLines,
+    this.dropCapPosition,
+  });
 
   @override
   Widget build(BuildContext context) {
+    final TextStyle? bodyLargeStyle = Theme.of(context).textTheme.bodyLarge;
     TextStyle textStyle = TextStyle(
-      color: Theme.of(context).textTheme.bodyText1?.color ?? Colors.black,
+      color: bodyLargeStyle?.color ?? Colors.black,
       fontSize: 14,
       height: 1,
-      fontFamily: Theme.of(context).textTheme.bodyText1?.fontFamily,
+      fontFamily: bodyLargeStyle?.fontFamily,
     ).merge(style);
 
     if (data == '') return Text(data, style: textStyle);
 
-    TextStyle capStyle = TextStyle(
+    final double textStyleFontSize = textStyle.fontSize!;
+
+    final TextStyle capStyle = TextStyle(
       color: textStyle.color,
-      fontSize: textStyle.fontSize! * 5.5,
+      fontSize: textStyleFontSize * 5.5,
       fontFamily: textStyle.fontFamily,
       fontWeight: textStyle.fontWeight,
       fontStyle: textStyle.fontStyle,
       height: 1,
     ).merge(dropCapStyle);
 
+    textStyle = textStyle.copyWith(
+        fontSize: MediaQuery.textScalerOf(context).scale(textStyleFontSize));
+
     double capWidth, capHeight;
     int dropCapChars = dropCap != null ? 0 : this.dropCapChars;
     CrossAxisAlignment sideCrossAxisAlignment = CrossAxisAlignment.start;
     MarkdownParser? mdData = parseInlineMarkdown ? MarkdownParser(data) : null;
-    final String dropCapStr = (mdData?.plainText ?? data).substring(0, dropCapChars);
+    final String dropCapStr =
+        (mdData?.plainText ?? data).substring(0, dropCapChars);
 
-    if (mode == DropCapMode.baseline && dropCap == null) return _buildBaseline(context, textStyle, capStyle);
+    if (mode == DropCapMode.baseline && dropCap == null) {
+      return _buildBaseline(context, textStyle, capStyle);
+    }
 
     // custom DropCap
     if (dropCap != null) {
@@ -119,7 +123,8 @@ class DropCapText extends StatelessWidget {
       capHeight = capPainter.height;
       if (forceNoDescent) {
         List<LineMetrics> ls = capPainter.computeLineMetrics();
-        capHeight -= ls.isNotEmpty ? ls[0].descent * 0.95 : capPainter.height * 0.2;
+        capHeight -=
+            ls.isNotEmpty ? ls[0].descent * 0.95 : capPainter.height * 0.2;
       }
     }
 
@@ -127,16 +132,18 @@ class DropCapText extends StatelessWidget {
     capWidth += dropCapPadding.left + dropCapPadding.right;
     capHeight += dropCapPadding.top + dropCapPadding.bottom;
 
-    MarkdownParser? mdRest = parseInlineMarkdown ? mdData!.subchars(dropCapChars) : null;
+    MarkdownParser? mdRest =
+        parseInlineMarkdown ? mdData!.subchars(dropCapChars) : null;
     String restData = data.substring(dropCapChars);
 
     TextSpan textSpan = TextSpan(
       text: parseInlineMarkdown ? null : restData,
       children: parseInlineMarkdown ? mdRest!.toTextSpanList() : null,
-      style: textStyle.apply(fontSizeFactor: MediaQuery.of(context).textScaleFactor),
+      style: textStyle,
     );
 
-    TextPainter textPainter = TextPainter(textDirection: textDirection, text: textSpan, textAlign: textAlign);
+    TextPainter textPainter = TextPainter(
+        textDirection: textDirection, text: textSpan, textAlign: textAlign);
     double lineHeight = textPainter.preferredLineHeight;
 
     int rows = ((capHeight - indentation.dy) / lineHeight).ceil();
@@ -148,7 +155,8 @@ class DropCapText extends StatelessWidget {
     }
 
     // BUILDER
-    return LayoutBuilder(builder: (BuildContext context, BoxConstraints constraints) {
+    return LayoutBuilder(
+        builder: (BuildContext context, BoxConstraints constraints) {
       double boundsWidth = constraints.maxWidth - capWidth;
       if (boundsWidth < 1) boundsWidth = 1;
 
@@ -158,7 +166,8 @@ class DropCapText extends StatelessWidget {
       if (rows > 0) {
         textPainter.layout(maxWidth: boundsWidth);
         double yPos = rows * lineHeight;
-        int charIndex = textPainter.getPositionForOffset(Offset(0, yPos)).offset;
+        int charIndex =
+            textPainter.getPositionForOffset(Offset(0, yPos)).offset;
         textPainter.maxLines = rows;
         textPainter.layout(maxWidth: boundsWidth);
         if (textPainter.didExceedMaxLines) charIndexEnd = charIndex;
@@ -175,9 +184,12 @@ class DropCapText extends StatelessWidget {
         children: <Widget>[
           // Text(totMillis.toString() + ' ms'),
           Row(
-            textDirection: dropCapPosition == null || dropCapPosition == DropCapPosition.start
+            textDirection: dropCapPosition == null ||
+                    dropCapPosition == DropCapPosition.start
                 ? textDirection
-                : (textDirection == TextDirection.ltr ? TextDirection.rtl : TextDirection.ltr),
+                : (textDirection == TextDirection.ltr
+                    ? TextDirection.rtl
+                    : TextDirection.ltr),
             crossAxisAlignment: sideCrossAxisAlignment,
             children: <Widget>[
               dropCap != null
@@ -199,10 +211,13 @@ class DropCapText extends StatelessWidget {
                 child: Container(
                   padding: EdgeInsets.only(top: indentation.dy),
                   width: boundsWidth,
-                  height:
-                      mode != DropCapMode.aside ? (lineHeight * min(maxLines ?? rows, rows)) + indentation.dy : null,
+                  height: mode != DropCapMode.aside
+                      ? (lineHeight * min(maxLines ?? rows, rows)) +
+                          indentation.dy
+                      : null,
                   child: RichText(
-                    overflow: (maxLines == null || (maxLines! > rows && overflow == TextOverflow.fade))
+                    overflow: (maxLines == null ||
+                            (maxLines! > rows && overflow == TextOverflow.fade))
                         ? TextOverflow.clip
                         : overflow,
                     maxLines: maxLines,
@@ -219,13 +234,19 @@ class DropCapText extends StatelessWidget {
               padding: EdgeInsets.only(left: indentation.dx),
               child: RichText(
                 overflow: overflow,
-                maxLines: maxLines != null && maxLines! > rows ? maxLines! - rows : null,
+                maxLines: maxLines != null && maxLines! > rows
+                    ? maxLines! - rows
+                    : null,
                 textAlign: textAlign,
                 textDirection: textDirection,
                 text: TextSpan(
-                  text: parseInlineMarkdown ? null : restData.substring(min(charIndexEnd, restData.length)),
-                  children: parseInlineMarkdown ? mdRest!.subchars(charIndexEnd).toTextSpanList() : null,
-                  style: textStyle.apply(fontSizeFactor: MediaQuery.of(context).textScaleFactor),
+                  text: parseInlineMarkdown
+                      ? null
+                      : restData.substring(min(charIndexEnd, restData.length)),
+                  children: parseInlineMarkdown
+                      ? mdRest!.subchars(charIndexEnd).toTextSpanList()
+                      : null,
+                  style: textStyle,
                 ),
               ),
             ),
@@ -234,7 +255,8 @@ class DropCapText extends StatelessWidget {
     });
   }
 
-  _buildBaseline(BuildContext context, TextStyle textStyle, TextStyle capStyle) {
+  RichText _buildBaseline(
+      BuildContext context, TextStyle textStyle, TextStyle capStyle) {
     MarkdownParser mdData = MarkdownParser(data);
 
     return RichText(
@@ -248,7 +270,7 @@ class DropCapText extends StatelessWidget {
           ),
           TextSpan(
             children: mdData.subchars(dropCapChars).toTextSpanList(),
-            style: textStyle.apply(fontSizeFactor: MediaQuery.of(context).textScaleFactor),
+            style: textStyle,
           ),
         ],
       ),
@@ -317,8 +339,10 @@ class MarkdownParser {
       final List<Markup> markups = [Markup(markup, isOpening)];
 
       if (bold && markup != MARKUP_BOLD) markups.add(Markup(MARKUP_BOLD, true));
-      if (italic && markup != MARKUP_ITALIC) markups.add(Markup(MARKUP_ITALIC, true));
-      if (underline && markup != MARKUP_UNDERLINE) markups.add(Markup(MARKUP_UNDERLINE, true));
+      if (italic && markup != MARKUP_ITALIC)
+        markups.add(Markup(MARKUP_ITALIC, true));
+      if (underline && markup != MARKUP_UNDERLINE)
+        markups.add(Markup(MARKUP_UNDERLINE, true));
 
       spans.add(
         MarkdownSpan(